+++ include/oo_alloc/IAllocator.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include <cstddef>
-#include <cstdint>
 
 namespace oo_alloc {
 
@@ -8,11 +7,11 @@ class IAllocator {
 public:
   virtual ~IAllocator() = default;
 
-  virtual void* alloc(std::size_t size, std::uint8_t align) = 0;
+  virtual void* alloc(std::size_t size, std::size_t align) = 0;
   virtual void free(void* ptr) = 0;
   virtual bool init(std::size_t size) = 0;
   virtual void clear() = 0;
-  virtual void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::uint8_t align) = 0;
+  virtual void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::size_t align) = 0;
 };
 
 }
