+++ include/oo_alloc/PoolAllocator.hpp
@@ -22,11 +22,11 @@ public:
   }
   ~PoolAllocator() override;
 
-  void* alloc(std::size_t size, std::uint8_t align) override;
+  void* alloc(std::size_t size, std::size_t align) override;
   void  free(void* ptr) override;
   bool  init(std::size_t size) override;
   void clear() override;
-  void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::uint8_t align) override;
+  void* realloc(void* ptr, std::size_t old_size, std::size_t new_size, std::size_t align) override;
 };
 
 }
